[#115] UI Enhancement
