+++ app/src/main/res/drawable/button_shape.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android" >
+    <item android:state_pressed="true" >
+        <shape android:shape="rectangle"  >
+            <corners android:radius="10dip" />
+            <stroke android:width="2dip" android:color="@color/colorPrimary" />
+            <gradient android:angle="-90" android:startColor="@color/colorBiright" android:endColor="@color/colorBiright" />
+        </shape>
+    </item>
+    <item android:state_focused="true">
+        <shape android:shape="rectangle"  >
+            <corners android:radius="10dip" />
+            <stroke android:width="1dip" android:color="@color/colorPrimary" />
+            <solid android:color="@color/colorPrimaryDark"/>
+            <gradient android:angle="-90" android:startColor="@color/colorBiright" android:endColor="@color/colorBiright" />
+
+        </shape>
+    </item>
+    <item >
+        <shape android:shape="rectangle"  >
+            <corners android:radius="10dip" />
+            <stroke android:width="2dip" android:color="@color/colorPrimary" />
+            <gradient android:angle="-90" android:startColor="@color/colorBiright" android:endColor="@color/colorBiright" />
+        </shape>
+    </item>
+</selector>
