+++ app/src/main/res/layout/export_dialog.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:orientation="vertical"
               android:layout_width="match_parent"
               android:layout_height="match_parent">
 
@@ -8,8 +7,9 @@
         android:id="@+id/export_name_title"
         android:layout_width="100dp"
         android:layout_height="50dp"
-        android:text="Name:"
-        android:gravity="center"
+        android:text="File Name:"
+        android:layout_marginStart="20dp"
+        android:gravity="start|center_vertical"
         android:textSize="20sp"
         android:layout_alignParentStart="true"
         android:layout_alignParentTop="true"/>
@@ -27,6 +27,7 @@
         android:orientation="horizontal"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:layout_marginStart="20dp"
         android:layout_centerVertical="true"
         android:layout_below="@id/export_name">
