+++ app/src/main/res/layout/save_dialog.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent">
+
+    <TextView
+        android:id="@+id/save_name_title"
+        android:layout_width="150dp"
+        android:layout_height="50dp"
+        android:text="Version Name:"
+        android:layout_marginStart="20dp"
+        android:gravity="start|center_vertical"
+        android:textSize="20sp"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentTop="true"/>
+
+    <EditText
+        android:id="@+id/save_name"
+        android:layout_width="match_parent"
+        android:layout_height="50dp"
+        android:textSize="20sp"
+        android:layout_alignParentTop="true"
+        android:layout_toEndOf="@id/save_name_title"/>
+
+</RelativeLayout>
