+++ app/src/main/res/menu/main.xml
@@ -1,12 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/action_export"
-        android:icon="@drawable/ic_export"
-        android:title="@string/feature_export"
-        app:showAsAction="ifRoom"/>
     <item
         android:id="@+id/action_save"
         android:icon="@android:drawable/ic_menu_save"
