+++ app/src/main/res/values/colors.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#3F51B5</color>
-    <color name="colorPrimaryDark">#303F9F</color>
-    <color name="colorAccent">#FF4081</color>
+    <color name="colorPrimary">#800009</color>
+    <color name="colorPrimaryDark">#800009</color>
+    <color name="colorAccent">#d57c81</color>
+    <color name="colorBiright">#f2e5e6</color>
 </resources>
