+++ app/src/main/res/values/strings.xml
@@ -5,7 +5,6 @@
     <string name="navigation_drawer_close">Close navigation drawer</string>
 
     <string name="feature_save">Save</string>
-    <string name="feature_save_name">Name</string>
     <string name="feature_save_cancel">Cancel</string>
     <string name="feature_save_success">Successfully saved!</string>
     <string name="feature_save_empty_name">Please input a name.</string>
